Fix type in test: avoid integer truncation.
This shouldn't impact the overall result of the test (it tends
to shrink to around 42-43, which is well under 256).

+++ test/test_theft_integration.c
@@ -1434,7 +1434,7 @@ TEST forking_hook(void) {
     PASS();
 }
 
-static size_t Fibonacci(uint8_t x) {
+static size_t Fibonacci(uint16_t x) {
     if (x < 2) {
         return 1;
     } else {
